Added astropy to extra requirements, moved fuctions from test_celestial to scriptfolder
